move AddCryptoAsset to accountController
